chore: refactor with clippy

+++ examples/control_circuits_demo/logger.rs
@@ -16,17 +16,17 @@ impl RerunLogger {
 
     pub fn input(&self, t: usize, input: f64) {
         self.rec.set_time_sequence("time", t as u32);
-        let _ = self.rec.log("val/input", &rerun::Scalar::new(input as f64));
+        let _ = self.rec.log("val/input", &rerun::Scalar::new(input));
     }
 
     pub fn output_pt1(&self, t: usize, output: f64) {
         self.rec.set_time_sequence("time", t as u32);
-        let _ = self.rec.log("val/pt1", &rerun::Scalar::new(output as f64));
+        let _ = self.rec.log("val/pt1", &rerun::Scalar::new(output));
     }
 
     pub fn output_pt2(&self, t: usize, output: f64) {
         self.rec.set_time_sequence("time", t as u32);
-        let _ = self.rec.log("val/pt2", &rerun::Scalar::new(output as f64));
+        let _ = self.rec.log("val/pt2", &rerun::Scalar::new(output));
     }
 }
